Finish the first version of the Landing page
This needs more work to do, I have to research about flex stretch, grow
and other features of flexbox to make the page responsive on mobile and table.

The use of margin, padding and heigh/width properties are also not the best practice,
so I have to work on that too.